# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EasyPlan Backend API",
    "description": "Intent-driven native task board API with JWT auth, HITL LangGraph checkpoints, and Async Queue SSE event streams.",
    "version": "1.3.0"
  },
  "paths": {
...
    "/api/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_api_tasks_get",
        "parameters": [
          {
            "name": "view_bucket",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "planned",
                    "my_day",
                    "backlog"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "View Bucket"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  },
                  "title": "Response List Tasks Api Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task",
        "operationId": "create_task_api_tasks_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/{task_id}/assist": {
      "post": {
        "tags": [
          "task-assist"
        ],
        "summary": "Start Task Assist",
        "operationId": "start_task_assist_api_tasks__task_id__assist_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskAssistRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAssistStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/assist/{request_id}": {
      "get": {
        "tags": [
          "task-assist"
        ],
        "summary": "Get Task Assist Snapshot",
        "operationId": "get_task_assist_snapshot_api_tasks__task_id__assist__request_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          },
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAssistRunSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "task-assist"
        ],
        "summary": "Cancel Task Assist",
        "operationId": "cancel_task_assist_api_tasks__task_id__assist__request_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          },
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAssistRunSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/assist/{request_id}/events": {
      "get": {
        "tags": [
          "task-assist"
        ],
        "summary": "Stream Task Assist Events",
        "description": "Task-assist-only SSE stream. The run_type is task_assist and allowed events are run_started, task_context_ready, assist_generation_started, assist_validation_started, still_running, assist_ready, done, and agent_error.",
        "operationId": "stream_task_assist_events_api_tasks__task_id__assist__request_id__events_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          },
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          },
          {
            "name": "last_event_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last Event Id"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task-assist SSE envelopes (one model per emitted event).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SseEventEnvelope"
                }
              },
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/assist/{request_id}/apply": {
      "post": {
        "tags": [
          "task-assist"
        ],
        "summary": "Apply Task Assist",
        "operationId": "apply_task_assist_api_tasks__task_id__assist__request_id__apply_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          },
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskAssistApplyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAssistApplyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskAssistApplyRequest": {
        "properties": {
          "selected_option_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Option Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TaskAssistApplyRequest"
      },
...
      "TaskAssistRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "start",
              "unstick",
              "decompose"
            ],
            "title": "Mode"
          },
          "user_context": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "User Context"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "mode"
        ],
        "title": "TaskAssistRequest"
      },
      "TaskAssistRunSnapshot": {
        "properties": {
          "task_id": {
            "type": "string",
            "format": "uuid",
            "title": "Task Id"
          },
          "thread_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Thread Id"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "start",
              "unstick",
              "decompose"
            ],
            "title": "Mode"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "ready",
              "applied",
              "cancelled",
              "failed",
              "expired"
            ],
            "title": "Status"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "queued",
                  "context_ready",
                  "generating",
                  "validating",
                  "ready",
                  "applied",
                  "cancelled",
                  "failed",
                  "expired"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "proposal": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/StartAssistProposal"
                  },
                  {
                    "$ref": "#/components/schemas/UnstickAssistProposal"
                  },
                  {
                    "$ref": "#/components/schemas/DecomposeAssistProposal"
                  }
                ],
                "discriminator": {
                  "propertyName": "proposal_type",
                  "mapping": {
                    "decompose": "#/components/schemas/DecomposeAssistProposal",
                    "start": "#/components/schemas/StartAssistProposal",
                    "unstick": "#/components/schemas/UnstickAssistProposal"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_id",
          "thread_id",
          "request_id",
          "mode",
          "status",
          "created_at",
          "updated_at",
          "expires_at"
        ],
        "title": "TaskAssistRunSnapshot"
      },
      "TaskAssistStartResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "format": "uuid",
            "title": "Task Id"
          },
          "thread_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Thread Id"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "start",
              "unstick",
              "decompose"
            ],
            "title": "Mode"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "ready",
              "applied",
              "cancelled",
              "failed",
              "expired"
            ],
            "title": "Status"
          },
          "events_url": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Events Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_id",
          "thread_id",
          "request_id",
          "mode",
          "status",
          "events_url"
        ],
        "title": "TaskAssistStartResponse"
      },
      "TaskCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "view_bucket": {
            "type": "string",
            "enum": [
              "planned",
              "my_day",
              "backlog"
            ],
            "title": "View Bucket",
            "default": "planned"
          },
          "is_in_my_day": {
            "type": "boolean",
            "title": "Is In My Day",
            "default": false
          },
          "parent_task_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Task Id"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TaskCreateRequest"
      },
...
      "TaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "parent_task_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Task Id"
          },
          "client_node_id": {
            "type": "string",
            "title": "Client Node Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "node_type": {
            "type": "string",
            "enum": [
              "group",
              "action"
            ],
            "title": "Node Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "view_bucket": {
            "type": "string",
            "title": "View Bucket"
          },
          "is_in_my_day": {
            "type": "boolean",
            "title": "Is In My Day"
          },
          "estimated_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Minutes"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "done_criteria": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Done Criteria"
          },
          "start_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Hint"
          },
          "fallback_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Action"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "phase_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase Id"
          },
          "phase_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase Order"
          },
          "practice_loop_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Practice Loop Id"
          },
          "assist_rollup": {
            "type": "boolean",
            "title": "Assist Rollup",
            "default": false
          },
          "assist_request_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assist Request Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "thread_id",
          "parent_task_id",
          "client_node_id",
          "title",
          "description",
          "node_type",
          "status",
          "view_bucket",
          "is_in_my_day",
          "estimated_minutes",
          "sort_order"
        ],
        "title": "TaskResponse"
      },
...
    }
  }
}